requirements.txt
aniso8601==7.0.0
attrs==24.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
Flask-GraphQL==2.0.1
Flask-SocketIO==5.4.1
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
grpcio==1.67.1
grpcio-tools==1.67.1
h11==0.14.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.3.0
MarkupSafe==3.0.2
platformdirs==4.3.6
promise==2.3
protobuf==5.28.3
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
Rx==1.6.3
setuptools==75.3.0
simple-websocket==1.1.0
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.1
wsproto==1.2.0
zeep==4.3.1
